Add npm script to run tests based on firefox configuraiton file
This script is being created even without the existance of such
configuration file for teaching purpose.
In another session (and branch) a refactor will happen, not only
to create the firefox config file, but also to create a file that
generates the basic configuration so that we don't have duplicated
code in different config files.